# postgres_execution_package/api_spec/openapi.yaml
openapi: 3.0.3
info:
  title: منصة قطع غيار السيارات - Auto Parts Marketplace API
  version: "1.9.0"
  description: >
    مواصفة API تنفيذية، تغطي 15 عملية (Operations) عبر 14 مسارًا (Paths):
    الواجهات التمثيلية الخمس الأصلية الموثّقة بالتفصيل في التصميم التفصيلي
    (الحزمة 2)، بالإضافة إلى 5 عمليات خدمة Auth (اعتماد Contract Extension &
    Implementation Plan): تسجيل خروج، عرض/إضافة/إزالة وسائل الهوية، وعرض
    مزوّدي الهوية المفعَّلين — فتصبح عمليات Auth تحديدًا 6 عمليات شاملة
    /auth/login الأصلية. وبالإضافة أيضًا إلى 5 عمليات خدمة PCT (الكتالوج):
    اقتراح قطعة، عرضها، اعتمادها (مدير النظام فقط، REQ-PCT-002)، إضافة اسم
    مترجَم، وتسجيل رقم OEM. وبالإضافة إلى 7 عمليات خدمة VCT (كتالوج
    السيارات): اقتراح واعتماد شركة مصنّعة (مدير النظام فقط، REQ-VCT-002)،
    عرضها، اقتراح موديل تحت شركة معتمَدة (REQ-VCT-003)، إضافة جيل، إضافة
    فئة سيارة، وعرض فئة سيارة. وبالإضافة إلى 4 عمليات خدمة CMP (التوافق):
    إنشاء سجل توافق بين قطعة معتمَدة وفئة صالحة (مدير النظام فقط،
    REQ-CMP-001/002)، عرض سجل، عرض كل سجلات قطعة، وأرشفته (مدير النظام
    فقط، REQ-CMP-003). وبالإضافة أيضًا إلى خدمتَي Store وInventory: إنشاء/
    عرض متجر، تغيير حالته (مدير نظام أو مشرف، REQ-STR-004)، نقل ملكيته
    (مدير نظام حصريًا، REQ-STR-006)؛ وعرض/تحديث كمية/تسعير/إخفاء/إظهار/
    أرشفة عنصر مخزون (البائع المالك فقط، REQ-STR-019) إلى جانب /inventory-items
    الأصلية (إنشاء العنصر، مع Idempotency-Key مطلوب). جميع مسارات هذا
...
    كل عملية موسومة بامتداد x-traceability
...
  x-governance: >
...
servers:
...
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        جلسة مستخدم صادرة عن وحدة IAM بعد تسجيل دخول ناجح (REQ-IAM-003)،
        مدعومة بجدول iam.sessions (CR-013). تُسلَّم عبر Set-Cookie بخصائص
        HttpOnly (لا يقرأها JavaScript) وSecure (HTTPS فقط) وSameSite=Lax
        كحد أدنى. يُخزَّن على الخادم بصمة SHA-256 للتوكن فقط، لا التوكن
        الخام (اتساقًا مع REQ-SEC-002). غير مطلوبة لعمليات الوصول العام مثل
        تسجيل الدخول نفسه أو البحث أو عرض مزوّدي الهوية المفعَّلين.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: >
        معرّف تتبع اختياري يرفقه المستهلك؛ إن غاب يُنشئ النظام معرّفًا جديدًا
        عند نقطة الدخول (SAD الحزمة 4، DD الحزمة 2 القسم 2.4).
      schema:
        type: string
        format: uuid
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        مفتاح عدم تكرار للعمليات الحساسة المعرَّضة لإعادة الإرسال
        (DD الحزمة 2، القسم 2.2). مطلوب فعليًا (required) على مستوى العملية
        عند الإنشاء، لا على مستوى الوثيقة العام.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      description: نموذج الخطأ الموحّد المعتمد لجميع واجهات API (DD الحزمة 2، القسم 2.1/2.2)
      required: [correlation_id, error_code, message]
      properties:
        correlation_id:
          type: string
          format: uuid
        error_code:
          type: string
          example: "PART_NOT_APPROVED"
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
...
    OfferSubmitRequest:
      type: object
      required: [amount, currency, provides_shipping]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          minLength: 3
          maxLength: 3
        provides_shipping:
          type: boolean
        notes:
          type: string
    OfferSubmitResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_code:
          type: string
        status:
          type: string
          enum: [submitted]
...
    PurchaseRequestCreateRequest:
      type: object
      required: [catalog_part_ref_id, trim_ref_id]
      properties:
        catalog_part_ref_id: { type: string, format: uuid }
        trim_ref_id: { type: string, format: uuid }
    PurchaseRequestResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        business_code: { type: string, nullable: true }
        buyer_user_ref_id: { type: string, format: uuid }
        catalog_part_ref_id: { type: string, format: uuid }
        trim_ref_id: { type: string, format: uuid }
        status: { type: string, enum: [open, under_review, fulfilled, expired, cancelled] }
    OfferResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        business_code: { type: string, nullable: true }
        purchase_request_id: { type: string, format: uuid }
        seller_store_ref_id: { type: string, format: uuid }
        amount: { type: number }
        currency: { type: string }
        provides_shipping: { type: boolean }
        notes: { type: string, nullable: true }
        status: { type: string, enum: [submitted, accepted, rejected, withdrawn, expired] }
    MessageSendRequest:
      type: object
      required: [context_type, context_ref_id, body]
      properties:
        context_type: { type: string, enum: [purchase_request, inventory_item] }
        context_ref_id: { type: string, format: uuid }
        body: { type: string, minLength: 1 }
    MessageResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        sender_user_ref_id: { type: string, format: uuid }
        body: { type: string }
        is_deleted_by_sender: { type: boolean }
        is_deleted_by_recipient: { type: boolean }
...
paths:
...
  /api/v1/purchase-requests:
    post:
      operationId: createPurchaseRequest
      summary: إنشاء طلب شراء لقطعة كتالوج معتمَدة (REQ-PUR-001/002)
      x-traceability: { requirements: [REQ-PUR-001, REQ-PUR-002], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CorrelationIdHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PurchaseRequestCreateRequest' }
      responses:
        '201': { description: أُنشئ الطلب بحالة open, content: { application/json: { schema: { $ref: '#/components/schemas/PurchaseRequestResponse' } } } }
        '409': { description: 'قطعة الكتالوج غير معتمدة', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/purchase-requests/{prId}:
    get:
      operationId: getPurchaseRequest
      summary: عرض طلب شراء بمعرّفه
      x-traceability: { requirements: [REQ-PUR-001], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: prId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: بيانات الطلب, content: { application/json: { schema: { $ref: '#/components/schemas/PurchaseRequestResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/purchase-requests/{prId}/cancel:
    post:
      operationId: cancelPurchaseRequest
      summary: إلغاء طلب شراء (المشتري صاحب الطلب حصرًا، REQ-PUR-009)
      x-traceability: { requirements: [REQ-PUR-009], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: prId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: أصبح ملغيًا, content: { application/json: { schema: { $ref: '#/components/schemas/PurchaseRequestResponse' } } } }
        '403': { description: ليس صاحب الطلب, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: انتقال حالة غير مسموح, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/offers/{offerId}/accept:
    post:
      operationId: acceptOffer
      summary: قبول عرض سعر واحد؛ رفض بقية العروض تلقائيًا (REQ-PUR-013/014)
      x-traceability: { requirements: [REQ-PUR-013, REQ-PUR-014], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: offerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: 'أصبح الطلب fulfilled', content: { application/json: { schema: { $ref: '#/components/schemas/PurchaseRequestResponse' } } } }
        '403': { description: ليس صاحب الطلب, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: حالة عرض غير صالحة للقبول, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/offers/{offerId}/withdraw:
    post:
      operationId: withdrawOffer
      summary: سحب عرض سعر قبل قبوله (البائع صاحب العرض حصرًا، REQ-PUR-018)
      x-traceability: { requirements: [REQ-PUR-018], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: offerId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: أصبح مسحوبًا, content: { application/json: { schema: { $ref: '#/components/schemas/OfferResponse' } } } }
        '403': { description: ليس صاحب العرض, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: غير موجود, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: 'العرض مقبول بالفعل؛ لا يجوز سحبه', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/messages:
    post:
      operationId: sendMessage
      summary: إرسال رسالة ضمن سياق محدَّد (REQ-COM-001/002)
      x-traceability: { requirements: [REQ-COM-001, REQ-COM-002], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CorrelationIdHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MessageSendRequest' }
      responses:
        '201': { description: أُرسلت الرسالة, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
        '400': { description: سياق غير معروف أو محتوى فارغ, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/v1/conversations/{conversationId}/messages:
    get:
      operationId: listConversationMessages
      summary: عرض رسائل محادثة (الظاهرة لصاحب الجلسة فقط، REQ-COM-007)
      x-traceability:
        requirements: [REQ-COM-002, REQ-COM-007]
        decisions: ["لا إنفاذ عضوية طرفين صارم بعد؛ Backlog موثَّق في الكود"]
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: قائمة الرسائل الظاهرة
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/MessageResponse' } }
  /api/v1/conversations/{conversationId}/messages/{messageId}:
    delete:
      operationId: deleteMessageForViewer
      summary: حذف نسبي لرسالة من منظور صاحب الجلسة فقط (لا حذف فعلي، REQ-COM-007)
      x-traceability: { requirements: [REQ-COM-007], decisions: [] }
      security: [{ sessionAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CorrelationIdHeader' }
        - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: messageId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: حُذفت نسبيًا, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }
        '404': { description: غير موجودة, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /api/v1/purchase-requests/{requestId}/offers:
    post:
      operationId: submitOffer
      summary: تقديم عرض سعر بائع على طلب شراء قائم
      x-traceability:
        requirements: [REQ-PUR-011, REQ-PUR-016]
        decisions: [CR-002]
      security: [{sessionAuth: []}]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferSubmitRequest'
      responses:
        '201':
          description: تم تقديم العرض بنجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferSubmitResponse'
        '409':
          description: طلب الشراء مغلق أو منتهي الصلاحية، أو يوجد عرض نشط سابق لنفس البائع
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: البائع غير مؤهَّل لهذا الطلب
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...